+++ docs/GUIDES.md
@@ -1,10 +1,8 @@
-## Implements Lib
+## Guides
 
-### DataSource
+### Release
 
-实现 `DataSource` 接口
-
-### EntityService
-
-实现 `EntityService` 接口
+标记 tag 之前需要更新 package version。
+通过 `git tag` 来标记版本，然后通过 `git push --tags` 来推送标记到远程仓库。
+GitHub workflow 会自动发布到 npm。
