+++ libs/component-angular/property/calculation/conditional-aggregation/conditional-aggregation.component.html
@@ -14,7 +14,6 @@
       [label]=" 'COMPONENTS.CALCULATION.VALUE' | translate: {Default: 'Value'} "
       [dataSettings]="dataSettings"
       [entityType]="entityType"
-      [coreService]="coreService"
       formControlName="value"
     >
     </ngm-parameter-select>
