+++ packages/core/package.json
@@ -1,12 +1,13 @@
 {
   "name": "@metad/ocap-core",
-  "version": "1.17.0-rc.1",
+  "version": "1.17.0-rc.2",
   "dependencies": {
     "tslib": "^2.3.0"
   },
   "main": "./src/index.js",
   "typings": "./src/index.d.ts",
   "peerDependencies": {
+    "@metad/store": "^1.17.0",
     "rxjs": ">=7.4.0",
     "uuid": "^8.3.2",
     "fast-deep-equal": "^3.1.3",
