+++ packages/core/src/lib/models/index.ts
@@ -8,3 +8,4 @@ export * from './indicator'
 export * from './query'
 export * from './dataset'
 export * from './filter-type'
+export * from './time'
