+++ packages/core/src/lib/services/smart-business.service.ts
@@ -1,7 +1,7 @@
 import { catchError, distinctUntilChanged, EMPTY, filter, Observable, shareReplay, switchMap } from 'rxjs'
 import { DSCoreService } from '../ds-core.service'
-import { FilterMergeMode, isTimeRangesSlicer, putFilter, workOutTimeRangeSlicers } from '../filter'
-import { getEntityProperty, QueryReturn } from '../models'
+import { FilterMergeMode, putFilter } from '../filter'
+import { getEntityProperty, isTimeRangesSlicer, QueryReturn, workOutTimeRangeSlicers } from '../models'
 import { FilteringLogic, IFilter, isAdvancedSlicer, ISlicer, isSlicer, QueryOptions } from '../types'
 import { isEmpty, isEqual, isString, nonNullable } from '../utils'
 import { EntityBusinessService, EntityBusinessState } from './entity-business.service'
