+++ packages/echarts/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@metad/ocap-echarts",
-  "version": "1.17.0-rc.1",
+  "version": "1.17.0-rc.2",
   "keywords": [
     "metad",
     "ocap",
