+++ packages/xmla/package.json
@@ -1,4 +1,4 @@
 {
   "name": "@metad/ocap-xmla",
-  "version": "1.17.0-rc.1"
+  "version": "1.17.0-rc.2"
 }
